requirements.txt
picamera
pillow
numpy
pyyaml
tensorflow

# + ffmpeg and imagemagick installed in $PATH
